chore: 파일 이름, 중복 코드 등 수정
